requirements.txt
fastapi[all]
uvicorn[standard]
pyzmq
pyais
bottle
minilogger
matplotlib
bitstring
pycryptodome